% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_logo.R
\name{ggplot_logo}
\alias{ggplot_logo}
\title{Add together two numbers}
\usage{
ggplot_logo(p, logo = system.file("extdata", "logo.png", package = "sothi"))
}
\arguments{
\item{p}{a ggplot}

\item{img}{path to an png-file}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers
}
\examples{
add(1, 1)
}
