% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_upk.R
\name{theme_upk}
\alias{theme_upk}
\title{Theme for ggplot UPK}
\usage{
theme_upk()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Theme for ggplot UPK
}
\examples{
add(1, 1)
}
