% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sothi-package.R
\docType{package}
\name{sothi-package}
\alias{sothi}
\alias{sothi-package}
\title{sothi: My own functions}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/socialthink/sothi}
  \item Report bugs at \url{https://github.com/socialthink/sothi/issues}
}

}
\author{
\strong{Maintainer}: Andreas Wyss \email{andreas.wyss@socialthink.ch} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
