% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sothi.R
\name{theme_sothi}
\alias{theme_sothi}
\title{Theme for ggplot}
\usage{
theme_sothi()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Theme for ggplot
}
\examples{
add(1, 1)
}
